Coordinates are now represented by long long ints.
Since long ints size is dependent on the architecture. Long long are instead
always 64bit (which is what we want to avoid overflows).

+++ src/point_cloud.h
@@ -2,7 +2,7 @@
 #define POINT_CLOUD_H
 
 typedef unsigned long cloud_size_t;
-typedef long coord_t;
+typedef long long coord_t;
 
 /**
  * Point, represented by a pair of coordinates
